Move tabletView to pages folder
