+++ front/src/pages/tabletview/components/OpenHoursSection.tsx
