+++ front/src/pages/tabletview/components/RangeOfficerStatusSection.tsx
