+++ front/src/pages/tabletview/rangeofficer.test.tsx
@@ -1,12 +1,12 @@
 import { render, screen, waitFor, fireEvent } from '@testing-library/react';
 import { act } from 'react-dom/test-utils';
 import Tabletview from './rangeofficer';
-import * as utils from '../utils/Utils';
-import testUtils from '../_TestUtils/TestUtils';
+import * as utils from '../../utils/Utils';
+import testUtils from '../../_TestUtils/TestUtils';
 import axios from 'axios';
 import { CookiesProvider } from 'react-cookie';
 import { BrowserRouter as Router } from 'react-router-dom';
-import InfoBox from '../infoBox/InfoBox';
+import InfoBox from '../../infoBox/InfoBox';
 import socketIOClient from 'socket.io-client';  // Import socket.io-client
 import { TestProviders } from '@/_TestUtils/TestProvides';
